% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{make_binomial_name}
\alias{make_binomial_name}
\title{Add the genus to create a binomial species name}
\usage{
make_binomial_name(genus_name, specific_name)
}
\arguments{
\item{genus_name}{The genus name, i.e. the first word of the species name in
binomial nomenclature.}

\item{specific_name}{The specific name, i.e. the second word of the species
name in binomial nomenclature.}
}
\value{
The binomial species name.
}
\description{
Add the genus to create a binomial species name
}
