% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{split_lineage}
\alias{split_lineage}
\title{Split lineages into taxa}
\usage{
split_lineage(lineage, pattern = "; ", ranks = taxonomic_ranks)
}
\arguments{
\item{lineage}{Character vector of taxonomic lineages.}

\item{pattern}{Pattern on which to split taxa in each lineage.}

\item{ranks}{Character vector of taxonomic ranks, used as column names in the
resultant data frame.}
}
\value{
A data frame of taxonomic assignments.
}
\description{
Split lineages into taxa
}
